+++ backend/src/main/java/com/playroute/playroute/entity/PlanEntity.java
@@ -2,6 +2,8 @@ package com.playroute.playroute.entity;
 
 import jakarta.persistence.*;
 import lombok.*;
+
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Map;
@@ -29,7 +31,7 @@ public class PlanEntity {
 
     private LocalDateTime endTime;
 
-    private Double totalCost;
+    private BigDecimal totalCost;
 
     @Type(value = JsonBinaryType.class)
     @Column(columnDefinition = "jsonb")
